# docs/swagger.yaml
basePath: /
definitions:
  model.User:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
      raiting:
        type: number
      status:
        type: string
    type: object
  service.UserSingIn:
    properties:
      password:
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
  service.UserSingUp:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - name
    - password
    - phone_number
    type: object
host: localhost:8080
info:
  contact:
    email: ripper@gmail.com
    name: API Support
  description: API for order taxi
  termsOfService: http://swagger.io/terms/
  title: InnoTaxi API
  version: "1.0"
paths:
  /users/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: refresh access token
      tags:
      - auth
  /users/auth/sing-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: phone number and password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.UserSingIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: user authentication
      tags:
      - auth
  /users/auth/sing-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: account info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.UserSingUp'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: registarte user
      tags:
      - auth
  /users/profile/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user's id to delete
        in: path
        name: id
        type: integer
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - Bearer: []
      summary: delete user
      tags:
      - user
    get:
      parameters:
      - description: user's id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - Bearer: []
      summary: get user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: rows to update
        in: body
        name: input
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: Unauthorized
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - Bearer: []
      summary: update user profile
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"